requirements.txt
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
Deprecated==1.2.18
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-pubsub==2.29.0
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.6.1
oauthlib==3.2.2
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
wrapt==1.17.2
zipp==3.22.0
